+++ bika/lims/content/abstractroutineanalysis.py
@@ -20,7 +20,7 @@ from bika.lims.content.abstractanalysis import AbstractAnalysis
 from bika.lims.content.abstractanalysis import schema
 from bika.lims.content.analysisspec import ResultsRangeDict
 from bika.lims.content.reflexrule import doReflexRuleAction
-from bika.lims.interfaces import IAnalysis, IRoutineAnalysis
+from bika.lims.interfaces import IAnalysis, IRoutineAnalysis, ICancellable
 from bika.lims.interfaces.analysis import IRequestAnalysis
 from bika.lims.workflow import doActionFor
 from bika.lims.workflow import getTransitionDate
@@ -114,7 +114,7 @@ schema = schema.copy() + Schema((
 
 
 class AbstractRoutineAnalysis(AbstractAnalysis):
-    implements(IAnalysis, IRequestAnalysis, IRoutineAnalysis)
+    implements(IAnalysis, IRequestAnalysis, IRoutineAnalysis, ICancellable)
     security = ClassSecurityInfo()
     displayContentsTab = False
     schema = schema