+++ bika/lims/interfaces/__init__.py
@@ -858,3 +858,9 @@ class IGetStickerTemplates(Interface):
 class IARReport(Interface):
     """Marker interface for published AR Reports
     """
+
+
+class ICancellable(Interface):
+    """Marker for those objects that can be cancelled, but their content
+    type is not bound to cancellation_workflow
+    """